Introduce a friendly repl.
When experimenting with, or demonstrating microstates, it helps to
have "live" microstates environment to work with. We've been using
runkit or codesandbox, but they each have their drawbacks, not least
of which is that they don't work without internet connectivity.

This introduces a repl script which boots a node.js repl server
containing all of the microstate functions attached to the global
namespace.

You can invoke it my running

```
$ yarn start
```

Some fixups were made to the class naming so that it is more easily
digestible:

1. Primitive types are just called primitive types now:
`Microstate<NumberType>` -> `Microstate<Number>`;
2. The `meta` property attached to each Microstate is now
non-enumerable, so that it doesn't show up in console.logs by default.
3. The `info` property of the `Id` type is now non-enumerable.